Setting TypeRegister::enable_experimental on the start of compile_syntax_node()
Slint has many nifty experimental features coming (like menu bars) that need to be enabled when compiling by setting.  While I was able to use this technique to experiment with menu bars, context menus did not function.  After some debugging, it appears that that the way that way experimental features is enabled is by modfying the `TypeRegister` object in a few ad hoc ways, along these lines:

````
dependency_registry.borrow_mut().expose_internal_types =
            is_builtin || state.borrow().tl.compiler_config.enable_experimental;
````

I do not quite understand why these enablements seem to be sufficient for `MenuBar` but not `ContextMenu` but in any case adding analogous logic to `compile_syntax_node()` does the trick.

All this said, the ideal solution seems (to me) to likely be adding a parameter to `TypeRegister::new()`, but that would be a much more intrusive change.